// crates/desktop/src/generated/dto/PromptBackupDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptBackupItemDto } from "./PromptBackupItemDto";

export type PromptBackupDto = {
	format: string;
	version: number;
	exported_at: number;
	prompts: Array<PromptBackupItemDto>;
};