// crates/desktop/src/generated/dto/PromptBackupItemDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PromptBackupItemDto = {
	id: string;
	title: string;
	description: string | null;
	category: string | null;
	content: string;
	tags: Array<string>;
	created_at: number;
	updated_at: number;
};